+++ sdk/dotnet/SecurityInsights/GetMetadata.cs
@@ -9,18 +9,18 @@ using Pulumi.Serialization;
 
 namespace Pulumi.AzureNative.SecurityInsights
 {
-    public static class GetMetadatum
+    public static class GetMetadata
     {
         /// <summary>
         /// Metadata resource definition.
         /// API Version: 2021-03-01-preview.
         /// </summary>
-        public static Task<GetMetadatumResult> InvokeAsync(GetMetadatumArgs args, InvokeOptions? options = null)
-            => Pulumi.Deployment.Instance.InvokeAsync<GetMetadatumResult>("azure-native:securityinsights:getMetadatum", args ?? new GetMetadatumArgs(), options.WithVersion());
+        public static Task<GetMetadataResult> InvokeAsync(GetMetadataArgs args, InvokeOptions? options = null)
+            => Pulumi.Deployment.Instance.InvokeAsync<GetMetadataResult>("azure-native:securityinsights:getMetadata", args ?? new GetMetadataArgs(), options.WithVersion());
     }
 
 
-    public sealed class GetMetadatumArgs : Pulumi.InvokeArgs
+    public sealed class GetMetadataArgs : Pulumi.InvokeArgs
     {
         /// <summary>
         /// The Metadata name.
@@ -46,14 +46,14 @@ namespace Pulumi.AzureNative.SecurityInsights
         [Input("workspaceName", required: true)]
         public string WorkspaceName { get; set; } = null!;
 
-        public GetMetadatumArgs()
+        public GetMetadataArgs()
         {
         }
     }
 
 
     [OutputType]
-    public sealed class GetMetadatumResult
+    public sealed class GetMetadataResult
     {
         /// <summary>
         /// The creator of the content item.
@@ -109,7 +109,7 @@ namespace Pulumi.AzureNative.SecurityInsights
         public readonly string? Version;
 
         [OutputConstructor]
-        private GetMetadatumResult(
+        private GetMetadataResult(
             Outputs.MetadataAuthorResponse? author,
 
             string contentId,