+++ sdk/nodejs/securityinsights/index.ts
@@ -38,7 +38,7 @@ export * from "./getIncidentComment";
 export * from "./getIncidentRelation";
 export * from "./getMCASDataConnector";
 export * from "./getMDATPDataConnector";
-export * from "./getMetadatum";
+export * from "./getMetadata";
 export * from "./getMicrosoftSecurityIncidentCreationAlertRule";
 export * from "./getOfficeDataConnector";
 export * from "./getProductSetting";
@@ -56,7 +56,7 @@ export * from "./incidentRelation";
 export * from "./listSourceControlRepositories";
 export * from "./mcasdataConnector";
 export * from "./mdatpdataConnector";
-export * from "./metadatum";
+export * from "./metadata";
 export * from "./microsoftSecurityIncidentCreationAlertRule";
 export * from "./officeDataConnector";
 export * from "./productSetting";
@@ -102,7 +102,7 @@ import { IncidentComment } from "./incidentComment";
 import { IncidentRelation } from "./incidentRelation";
 import { MCASDataConnector } from "./mcasdataConnector";
 import { MDATPDataConnector } from "./mdatpdataConnector";
-import { Metadatum } from "./metadatum";
+import { Metadata } from "./metadata";
 import { MicrosoftSecurityIncidentCreationAlertRule } from "./microsoftSecurityIncidentCreationAlertRule";
 import { OfficeDataConnector } from "./officeDataConnector";
 import { ProductSetting } from "./productSetting";
@@ -155,8 +155,8 @@ const _module = {
                 return new MCASDataConnector(name, <any>undefined, { urn })
             case "azure-native:securityinsights:MDATPDataConnector":
                 return new MDATPDataConnector(name, <any>undefined, { urn })
-            case "azure-native:securityinsights:Metadatum":
-                return new Metadatum(name, <any>undefined, { urn })
+            case "azure-native:securityinsights:Metadata":
+                return new Metadata(name, <any>undefined, { urn })
             case "azure-native:securityinsights:MicrosoftSecurityIncidentCreationAlertRule":
                 return new MicrosoftSecurityIncidentCreationAlertRule(name, <any>undefined, { urn })
             case "azure-native:securityinsights:OfficeDataConnector":