+++ sdk/nodejs/securityinsights/v20210301preview/getMetadata.ts
@@ -8,7 +8,7 @@ import * as utilities from "../../utilities";
 /**
  * Metadata resource definition.
  */
-export function getMetadatum(args: GetMetadatumArgs, opts?: pulumi.InvokeOptions): Promise<GetMetadatumResult> {
+export function getMetadata(args: GetMetadataArgs, opts?: pulumi.InvokeOptions): Promise<GetMetadataResult> {
     if (!opts) {
         opts = {}
     }
@@ -16,7 +16,7 @@ export function getMetadatum(args: GetMetadatumArgs, opts?: pulumi.InvokeOptions
     if (!opts.version) {
         opts.version = utilities.getVersion();
     }
-    return pulumi.runtime.invoke("azure-native:securityinsights/v20210301preview:getMetadatum", {
+    return pulumi.runtime.invoke("azure-native:securityinsights/v20210301preview:getMetadata", {
         "metadataName": args.metadataName,
         "operationalInsightsResourceProvider": args.operationalInsightsResourceProvider,
         "resourceGroupName": args.resourceGroupName,
@@ -24,7 +24,7 @@ export function getMetadatum(args: GetMetadatumArgs, opts?: pulumi.InvokeOptions
     }, opts);
 }
 
-export interface GetMetadatumArgs {
+export interface GetMetadataArgs {
     /**
      * The Metadata name.
      */
@@ -46,7 +46,7 @@ export interface GetMetadatumArgs {
 /**
  * Metadata resource definition.
  */
-export interface GetMetadatumResult {
+export interface GetMetadataResult {
     /**
      * The creator of the content item.
      */